chore: Add space props on button menu
