+++ packages/pancake-uikit/src/components/ButtonMenu/ButtonMenu.tsx
@@ -1,7 +1,33 @@
 import React, { cloneElement, Children, ReactElement } from "react";
-import StyledButtonMenu from "./StyledButtonMenu";
+import styled, { DefaultTheme } from "styled-components";
+import { space } from "styled-system";
 import { scales, variants } from "../Button/types";
 import { ButtonMenuProps } from "./types";
+
+interface StyledButtonMenuProps extends ButtonMenuProps {
+  theme: DefaultTheme;
+}
+
+const getBackgroundColor = ({ theme, variant }: StyledButtonMenuProps) => {
+  return theme.colors[variant === variants.SUBTLE ? "input" : "tertiary"];
+};
+
+const getBorderColor = ({ theme, variant }: StyledButtonMenuProps) => {
+  return theme.colors[variant === variants.SUBTLE ? "inputSecondary" : "disabled"];
+};
+
+const StyledButtonMenu = styled.div<StyledButtonMenuProps>`
+  background-color: ${getBackgroundColor};
+  border-radius: 16px;
+  display: inline-flex;
+  border: 1px solid ${getBorderColor};
+
+  & > button + button,
+  & > a + a {
+    margin-left: 2px; // To avoid focus shadow overlap
+  }
+  ${space}
+`;
 
 const ButtonMenu: React.FC<ButtonMenuProps> = ({
   activeIndex = 0,
@@ -9,9 +35,10 @@ const ButtonMenu: React.FC<ButtonMenuProps> = ({
   variant = variants.PRIMARY,
   onItemClick,
   children,
+  ...props
 }) => {
   return (
-    <StyledButtonMenu variant={variant}>
+    <StyledButtonMenu variant={variant} {...props}>
       {Children.map(children, (child: ReactElement, index) => {
         return cloneElement(child, {
           isActive: activeIndex === index,
