--- packages/pancake-uikit/src/components/ButtonMenu/StyledButtonMenu.tsx
@@ -1,29 +1,0 @@
-import styled, { DefaultTheme } from "styled-components";
-import { Variant, variants } from "../Button/types";
-
-type StyledButtonMenuProps = {
-  variant: Variant;
-  theme: DefaultTheme;
-};
-
-const getBackgroundColor = ({ theme, variant }: StyledButtonMenuProps) => {
-  return theme.colors[variant === variants.SUBTLE ? "input" : "tertiary"];
-};
-
-const getBorderColor = ({ theme, variant }: StyledButtonMenuProps) => {
-  return theme.colors[variant === variants.SUBTLE ? "inputSecondary" : "disabled"];
-};
-
-const StyledButtonMenu = styled.div<{ variant: Variant }>`
-  background-color: ${getBackgroundColor};
-  border-radius: 16px;
-  display: inline-flex;
-  border: 1px solid ${getBorderColor};
-
-  & > button + button,
-  & > a + a {
-    margin-left: 2px; // To avoid focus shadow overlap
-  }
-`;
-
-export default StyledButtonMenu;
