+++ packages/pancake-uikit/src/components/ButtonMenu/types.ts
@@ -1,10 +1,11 @@
+import { SpaceProps } from "styled-system";
 import { BaseButtonProps, Scale, variants } from "../Button/types";
 
 export interface ButtonMenuItemProps extends BaseButtonProps {
   isActive?: boolean;
 }
 
-export interface ButtonMenuProps {
+export interface ButtonMenuProps extends SpaceProps {
   variant?: typeof variants.PRIMARY | typeof variants.SUBTLE;
   activeIndex?: number;
   onItemClick?: (index: number) => void;
